chore(localdebug): rename action name tools/install to prerequisite/install (#7152)
* chore(localdebug): rename tools/install to prerequisite/install

* chore: rename message name

* refactor(localdebug): update folder name from tools to prerequisite

* chore: rename tools/install in debug migration
