+++ packages/fx-core/resource/package.nls.json
@@ -1118,21 +1118,21 @@
   "driver.spfx.warn.noTenantAppCatalogFound": "No tenant app catalog found, retry: %s",
   "driver.spfx.error.failedToGetAppCatalog": "Cannot get app catalog site url after creation. You may need wait a few minutes and retry.",
   "driver.spfx.error.noValidAppCatelog": "There is no valid app catalog in your tenant. You can update the property 'createAppCatalogIfNotExist' in app.yml to true if you want us to create it for you or you can create it by yourself.",
-  "driver.tools.error.invalidParameter": "Following parameter is missing or invalid for %s action: %s.",
-  "driver.tools.error.funcInstallationError": "Failed to check and install Azure Functions Core Tools.",
-  "driver.tools.error.dotnetInstallationError": "Failed to check and install .NET Core SDK.",
-  "driver.tools.description": "installing dependencies",
-  "driver.tools.progressBar.title": "Driver tools/install",
-  "driver.tools.progressBar.devCert": "Checking and generating the development certificate for localhost.",
-  "driver.tools.progressBar.dotnet": "Checking and installing .NET Core SDK.",
-  "driver.tools.progressBar.func": "Checking and installing Azure Functions Core Tools.",
-  "driver.tools.summary.devCert.trusted.succuss": "Development certificate for localhost is installed.",
-  "driver.tools.summary.devCert.notTrusted.succuss": "Development certificate for localhost is generated.",
-  "driver.tools.summary.devCert.skipped": "Skip trusting development certificate for localhost.",
-  "driver.tools.summary.func.installedWithPath": "Azure Functions Core Tools is installed at %s.",
-  "driver.tools.summary.func.installed": "Azure Functions Core Tools is installed.",
-  "driver.tools.summary.dotnet.installedWithPath": ".NET Core SDK is installed at %s.",
-  "driver.tools.summary.dotnet.installed": ".NET Core SDK is installed.",
+  "driver.prerequisite.error.invalidParameter": "Following parameter is missing or invalid for %s action: %s.",
+  "driver.prerequisite.error.funcInstallationError": "Failed to check and install Azure Functions Core Tools.",
+  "driver.prerequisite.error.dotnetInstallationError": "Failed to check and install .NET Core SDK.",
+  "driver.prerequisite.description": "installing dependencies",
+  "driver.prerequisite.progressBar.title": "Driver prerequisite/install",
+  "driver.prerequisite.progressBar.devCert": "Checking and generating the development certificate for localhost.",
+  "driver.prerequisite.progressBar.dotnet": "Checking and installing .NET Core SDK.",
+  "driver.prerequisite.progressBar.func": "Checking and installing Azure Functions Core Tools.",
+  "driver.prerequisite.summary.devCert.trusted.succuss": "Development certificate for localhost is installed.",
+  "driver.prerequisite.summary.devCert.notTrusted.succuss": "Development certificate for localhost is generated.",
+  "driver.prerequisite.summary.devCert.skipped": "Skip trusting development certificate for localhost.",
+  "driver.prerequisite.summary.func.installedWithPath": "Azure Functions Core Tools is installed at %s.",
+  "driver.prerequisite.summary.func.installed": "Azure Functions Core Tools is installed.",
+  "driver.prerequisite.summary.dotnet.installedWithPath": ".NET Core SDK is installed at %s.",
+  "driver.prerequisite.summary.dotnet.installed": ".NET Core SDK is installed.",
   "driver.file.error.invalidParameter": "Following parameter is missing or invalid for %s action: %s.",
   "driver.file.error.unhandledError": "Unhandled error happened in %s action: %s",
   "driver.file.description": "generates environment variables to .env file",
