+++ templates/scenarios/js/sso-tab/teamsfx/app.local.yml.tpl
@@ -48,7 +48,7 @@ configureApp:
     # TEAMS_APP_ID: the id of Teams app
 
 deploy:
-  - uses: tools/install # Install dependencies
+  - uses: prerequisite/install # Install dependencies
     with:
       devCert:
         trust: true
