+++ templates/scenarios/ts/non-sso-tab/teamsfx/app.local.yml
@@ -28,7 +28,7 @@ configureApp:
     # TEAMS_APP_ID: the id of Teams app
 
 deploy:
-  - uses: tools/install # Install dependencies
+  - uses: prerequisite/install # Install dependencies
     with:
       devCert:
         trust: true
